pyproject.toml
[project]
name = "teamsheets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "gspread>=6.0.0",
    "google-auth>=2.25.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    "pillow>=10.0.0",
    "gunicorn>=21.0.0",
    "python-dotenv>=1.0.0",
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.2",
    "requests>=2.32.5",
]
